Moved intern.models.InternGroup to core.models.Group to fix #78
